+++ R/plotExperiment.R
@@ -1,6 +1,6 @@
 #' Plots the output of  \code{\link{runExperiment}}.
 #'
-#' @description It takes the output of \code{\link{runExperiment}}, and generates plots of ecological memory patterns for a large number of simulated pollen curves.
+#' @description Takes the output of \code{\link{runExperiment}}, and generates plots of ecological memory patterns for a large number of simulated pollen curves.
 #'
 #'
 #'@usage plotExperiment(
@@ -23,7 +23,7 @@
 #' @param experiment.output list, output of  \code{\link{runExperiment}}.
 #' @param parameters.file dataframe of simulation parameters.
 #' @param experiment.title character string, title of the plot.
-#' @param sampling.names vector of character strings with the names of the columns used in the argument \code{simulations.file} of \code{\link{runExperiment}}. If \code{NULL}, no pdf plot is produced.
+#' @param sampling.names vector of character strings with the names of the columns used in the argument \code{simulations.file} of \code{\link{runExperiment}}.
 #' @param filename character string, path and name (without extension) of the output pdf file.
 #' @param legend.position legend position in ggplot object. One of "bottom", "right", "none".
 #' @param R2 boolean. If \code{TRUE}, pseudo R-squared values are printed along with the traits of the virtual taxa.
@@ -60,7 +60,7 @@ plotExperiment <- function(experiment.output = NULL,
 
   #output of experiment.output to long table
   simulation.df <- experimentToTable(experiment.output = experiment.output,
-                                    parameters.file = parameters,
+                                    parameters.file = parameters.file,
                                     sampling.names = sampling.names,
                                     R2 = R2)