+++ R/plotInteraction.R
@@ -16,7 +16,7 @@
 #' @param data dataframe used to fit the model.
 #' @param x character string, name of column in \code{data} to be plotted in the x axis.
 #' @param y character string, name of column in \code{data} to be plotted in the y axis.
-#' @param z character string, name of column in \code{data} to be plotted as a surface. Generally, the name of the response variable in \code{model}.
+#' @param z character string, name of the response variable column in \code{data}, plotted as the surface.
 #' @param grid numeric, resolution of the x and y axes.
 #' @param point.size.range numeric vector with two values defining the range size of the points representing the observed data.
 #'
@@ -56,14 +56,14 @@ plotInteraction <- function(model = NULL,
   }
 
   #predicting different types of models
-  if(class(model)=="ranger"){
+  if(inherits(model, "ranger")){
     newdata$prediction <- predict(model, newdata)$predictions
   }
-  if(class(model)=="rpart"){
+  if(inherits(model, "rpart")){
     newdata$prediction <- predict(model, newdata, type="vector")
 
   }
-  if(!(class(model) %in% c("ranger", "rpart"))){
+  if(!inherits(model, c("ranger", "rpart"))){
     newdata$prediction <- predict(model, newdata, type="response")
   }